(fix): renamed smart request file
